+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - New `GraphTool.createLabeledGraph(...)` method to create labeled plot
 - New `PluginInitializer.functionalities()` method to load functionality information
 - New `Regression.getErrorSquare()` method
+- New `Outcome.BLOCKED` and `Outcome.SKIPPED` test result values
 - New `description` field in `StepElement`, `TestElement`, `SpecElement` and `TestRun` classes
 - New `TestRun.responsibleParty` field which can hold the responsible person/organization
 - New `PageObject.pageObject(...)` and `StepLibrary.pageObject(...)` methods that return a page object of the specified type